requirements.txt
azure_storage==0.37.0
fitz==0.0.1.dev2
langchain==0.1.20
langchain_community==0.0.38
langchain_core==0.1.52
langchain_openai==0.1.6
langchain_text_splitters==0.0.1
openai==1.28.1
pandas==2.2.2
pdfplumber==0.11.0
python-dotenv==1.0.1
qdrant_client==1.9.1
streamlit==1.33.0
